+++ apis/conversion/src/hypercube/index.js
@@ -1,0 +1,7 @@
+import exportProperties from './export-properties';
+import importProperties from './import-properties';
+
+export default {
+  exportProperties,
+  importProperties,
+};
